+++ crates/block-producer/src/lib.rs
@@ -15,6 +15,9 @@ pub mod errors;
 #[cfg(not(feature = "testing"))]
 mod errors;
 
+#[cfg(test)]
+mod errors_test;
+
 pub mod server;
 pub use server::BlockProducer;